requirements.txt

pandas>=2.2
duckdb>=1.0
pyarrow>=16.0
openpyxl>=3.1
plotly>=5.22
streamlit>=1.36
typer[all]>=0.12.3
pydantic>=2.7
rich>=13.7
ulid-py>=1.1
PyYAML>=6.0
pytest>=8.2
ruff>=0.5
